📆 Add the android calendar adapter (#4492)
* :bug: Fix android calendar service time col reading type

* :construction: Simplify and fix calendar layout manager offset calculation

* :construction: Make the anchored data immutable again

* :construction: Use ISubject instead of Subject

* :construction: Make the Calendar entry vh inherit from the base vh

* :sparkles: Add the layout anchor calculator

* :construction: Remove test data and add basic touch handling

* :construction: Bind calendar items to adapter

* :construction: Bind calendar item taps to adapter item touches

* :art: Update calendar entry cell layout

* :construction: Adapt anchor calculation to consider two columns

* :construction: Add observable and backing field to check for linked calendars

* :construction: Update bindings to consider two columns layout

* :lipstick:

* :construction: Add diffutil to anchors layout calculation

* :construction: Remove anchors width to simplify touch handling

* :construction: Fix simple item in edit mode handling

* :construction: Add touch on empty space to create te handling on rv

* :fire:

* :construction: Observe and subscribe to changes in the background

* :construction: Fix thread selection when running updates in the calendar adapter

* :construction: Fix calendar items update binding

* :fire:

* :fire:

* :white_check_mark:

+++ Toggl.Giskard/Services/CalendarServiceAndroid.cs
@@ -104,8 +104,8 @@ namespace Toggl.Giskard.Services
         private static CalendarItem calendarItemFromCursor(ICursor cursor)
         {
             var id = cursor.GetString(eventIdIndex);
-            var startDateUnixTime = cursor.GetInt(eventStartDateIndex);
-            var endDateUnixTime = cursor.GetInt(eventEndDateIndex);
+            var startDateUnixTime = cursor.GetLong(eventStartDateIndex);
+            var endDateUnixTime = cursor.GetLong(eventEndDateIndex);
             var description = cursor.GetString(eventDescriptionIndex);
             var color = cursor.GetString(eventColorIndex);
             var calendarId = cursor.GetString(eventCalendarIdIndex);
